Task times measured in doubles, not ints
This really impacts millisecond timing and is probably resulting in much
larger errors during logging.

+++ source/MrsWatson.c
@@ -95,7 +95,7 @@ int mrsWatsonMain(ErrorReporter errorReporter, int argc, char** argv) {
   boolByte finishedReading = false;
   int hostTaskId;
   SampleSource silentSampleInput;
-  unsigned long totalProcessingTime = 0;
+  double totalProcessingTime = 0.0;
   unsigned long stopFrame;
   double timePercentage;
   int i;
